topPlayers_componentizado
