+++ src/components/Home/Home.scss
@@ -4,45 +4,6 @@
     background: #0F5156;
 }
 
-.top-three_container {
-    width: 90%;
-    height: 380px;
-    margin: auto;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: space-evenly;
-    background: linear-gradient(180deg, #C65F72 0%, #2A855A 50%, #0F5156 100%);
-    border-radius: 7px;
-    position: relative;
-}
-
-.top-three_container h2 {
-    position: absolute;
-    top: 15px;
-    margin: auto;
-    font-family: 'Kanit';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 53.3333px;
-    line-height: 80px;
-    color: #FFFFFF;
-    text-shadow: 0px 0px 6px rgba(0, 0, 0, 0.49);
-}
-
-.top-ten_container {
-    width: 70%;
-    height: 480px;
-    margin: auto;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: space-evenly;
-    background: #0F5156;
-}
-
 #home-wallpaper{
     width: 100%;
     z-index: 1;
