+++ src/components/Player/Player.jsx
@@ -3,13 +3,13 @@ import './Player.scss'
 const Player = ({ position, name, img, ranking, alt, size }) => {
     return (
         <div className={`playerCard_container ${size && 'small'}`}>
-            <div className={`playerCard_img_container ${size && 'small'}`}>
+            <div className={`playerCard_img_container ${size && 'small'} ${position === 1 && 'one' || position === 2 && 'two' ||  position === 3 && 'three'}`}>
                 <img src={img} alt={`${name}_${alt}`} />
             </div>
             <div className={`playerCard_text_container ${size && 'small'}`}>
-                <h4>{position} position</h4>
-                <h3>{name}</h3>
-                <p>{ranking} puntos</p>
+                <h4 className={`playerCard_text_position ${size && 'small'} ${position === 1 && 'p-one' || position === 2 && 'p-two' ||  position === 3 && 'p-three'}`}>{position}° Puesto</h4>
+                <h3 className={`playerCard_text_name ${size && 'small'}`}>{name}</h3>
+                <p className={`playerCard_text_ranking ${size && 'small'}`}>{ranking} puntos</p>
             </div>
         </div>
     )
