+++ src/components/Player/Player.scss
@@ -1,62 +1,41 @@
 
 .playerCard_container {
-    width: 360px;
+    width: 260px;
     height: 170px;
     display: flex;
     flex-direction: row;
     align-items: center;
     justify-content: space-around;
 }
+
 .playerCard_container.small {
-    width: 276px;
+    width: 236px;
     height: 128px;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-around;
 }
 
 .playerCard_img_container {
-    width: 170px;
-    height: 100%;
+    width: 120px;
+    height: 120px;
     border-radius: 50%;
     display: flex;
     align-items: center;
     justify-content: center;
-    background: #6DCC5E;
-}
-.playerCard_img_container.small {
-    width: 126px;
-    height: 126px;
     background: #2A7F56;
 }
 
 .playerCard_img_container img {
-    width: 140px;
-    height: 140px;
-}
-.playerCard_img_container.small img {
-    width: 104px;
-    height: 104px;
+    width: 100px;
+    height: 100px;
 }
 
-.playerCard_text_container h4 {
+.playerCard_text_position {
     color:#6DCC5E;
     font-weight: 700;
-    font-size: 26px;
-}
-.playerCard_text_container.small h4 {
-    color:#6DCC5E;
-    font-weight: 700;
-    font-size: 18px;
+    font-size: 22px;
 }
 
-.playerCard_text_container h3 {
-    font-weight: 400;
-    font-size: 40px;
-    color: #fff;
-}
-.playerCard_text_container.small h3 {
+.playerCard_text_name {
+    text-align: left;
     font-weight: 400;
     font-size: 32px;
     color: #fff;
@@ -64,11 +43,29 @@
 
 .playerCard_text_container p {
     font-weight: 400;
-    font-size: 26px;
+    font-size: 22px;
     color: #fff;
 }
-.playerCard_text_container.small p {
-    font-weight: 400;
-    font-size: 18px;
-    color: #fff;
+
+.one {
+    background: #FED881;
+}
+.two {
+    background: #B9DF67;
+}
+.three {
+    background: #6DCC5E;
+}
+.p-one {
+    color: #FED881;
+}
+.p-two {
+    color: #B9DF67;
+}
+.p-three {
+    color: #6DCC5E;
+}
+
+.small {
+    transform: scale(90%,90%);
 }
