+++ src/components/TopPlayers/TopPlayers.scss
@@ -1,0 +1,54 @@
+.top-players {
+    width: 100%;
+    height: auto;
+}
+
+.top-players_container {
+    width: 90%;
+    height: auto;
+    text-align: center;
+    margin: auto;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.top-three {
+    width: 100%;
+    height: 300px;
+    margin: auto;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background: linear-gradient(180deg, #C65F72 0%, #2A855A 50%, #0F5156 100%);
+    border-radius: 7px;
+    position: relative;
+}
+
+.top-three-title {
+    color: #fff;
+    margin-top: 30px;
+}
+
+.top-three_container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    align-items: center;
+    justify-content: space-evenly;
+}
+
+.top-ten_container {
+    width: 90%;
+    height: 300px;
+    margin: auto;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+    background: #0F5156;
+}
